Resolve circular dependency between Entity and Registry
