+++ script/Modules/Core/Components/ChildrenComponent.lua
@@ -1,7 +1,7 @@
 local Component = require("Core.ECS.Component")
 
 ---@class ChildrenComponent: Component
----@field children EntityId[]
+---@field children Entity[]
 ---@overload fun(self: ChildrenComponent) : ChildrenComponent subclass internal
 ---@overload fun(): ChildrenComponent subclass external
 local ChildrenComponent = Subclass("ChildrenComponent", Component, function(self)
@@ -10,22 +10,22 @@ local ChildrenComponent = Subclass("ChildrenComponent", Component, function(self
     self.children = {}
 end)
 
----@param entityId EntityId
-function ChildrenComponent:addChild(entityId)
-    insert(self.children, entityId)
+---@param entity Entity
+function ChildrenComponent:addChild(entity)
+    insert(self.children, entity)
 end
 
----@param entityId EntityId
-function ChildrenComponent:removeChild(entityId)
+---@param entity Entity
+function ChildrenComponent:removeChild(entity)
     for i, v in ipairs(self.children) do
-        if v == entityId then
+        if v == entity then
             remove(self.children, i)
             break
         end
     end
 end
 
----@return Iterator<EntityId>
+---@return Iterator<Entity>
 function ChildrenComponent:iterChildren()
     return Iterator(self.children)
 end
