+++ script/Modules/Core/Components/ParentComponent.lua
@@ -1,21 +1,21 @@
 local Component = require("Core.ECS.Component")
 
 ---@class ParentComponent: Component
----@field parent EntityId
----@overload fun(self: ParentComponent, parent: EntityId) : ParentComponent subclass internal
----@overload fun(parent: EntityId): ParentComponent subclass external
+---@field parent Entity
+---@overload fun(self: ParentComponent, parent: Entity) : ParentComponent subclass internal
+---@overload fun(parent: Entity): ParentComponent subclass external
 local ParentComponent = Subclass("ParentComponent", Component, function(self, parent)
     self:setComponentName("Parent")
 
     self.parent = parent
 end)
 
----@param parent EntityId
+---@param parent Entity
 function ParentComponent:setParent(parent)
     self.parent = parent
 end
 
----@return EntityId
+---@return Entity
 function ParentComponent:getParent()
     return self.parent
 end
